# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GoShort API
  version: 0.5.0
  description: |
    Self-hosted URL shortener API.

    **Auth:** All write endpoints require `X-API-Key` header (Phase 3).
    Public endpoints: `GET /{code}`, `GET /health`, `GET /metrics`, `GET /docs`.
servers:
  - url: http://localhost:8080
    description: Local development
security: []
paths:
  /api/v1/urls:
    post:
      summary: Create short URL
      operationId: createURL
      tags: [URLs]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateURLRequest"
            example:
              url: "https://example.com/very/long/path?with=params"
              custom_alias: "my-link"
              expires_in: "30d"
      responses:
        "201":
          description: Short URL created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateURLResponse"
              example:
                short_code: "my-link"
                short_url: "http://localhost:8080/my-link"
                original_url: "https://example.com/very/long/path?with=params"
                expires_at: "2025-07-01T00:00:00Z"
                created_at: "2025-06-01T12:00:00Z"
        "400":
          description: Invalid URL or expires_in format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_url:
                  value:
                    error:
                      code: invalid_url
                      message: The URL format is invalid
                invalid_expires:
                  value:
                    error:
                      code: invalid_expires
                      message: The expires_in duration is invalid
        "409":
          description: Alias already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: alias_taken
                  message: The requested alias is already in use
        "422":
          description: Alias validation failed, reserved path, or URL flagged as unsafe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_alias:
                  value:
                    error:
                      code: invalid_alias
                      message: The alias format is invalid
                reserved_path:
                  value:
                    error:
                      code: reserved_path
                      message: The alias is a reserved path
                unsafe_url:
                  value:
                    error:
                      code: unsafe_url
                      message: The URL has been flagged as unsafe
        "429":
          description: Too many requests
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: string
              example: "60"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.

    get:
      summary: List URLs (paginated)
      operationId: listURLs
      tags: [URLs]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          example: 20
      responses:
        "200":
          description: Paginated URL list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListURLsResponse"
              example:
                data:
                  - short_code: k7Xm2p
                    short_url: "http://localhost:8080/k7Xm2p"
                    original_url: "https://example.com"
                    is_custom: false
                    expires_at: null
                    created_at: "2025-06-01T12:00:00Z"
                    click_count: 42
                pagination:
                  page: 1
                  per_page: 20
                  total: 156
                  total_pages: 8
        "429":
          description: Too many requests
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: string
              example: "60"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.
...
  /api/v1/urls/public:
    post:
      summary: Create short URL (public)
      operationId: publicCreateURL
      tags: [URLs]
      description: |
        No authentication required. Forces 30-day expiry. No custom alias allowed.
        Rate limited to 5 requests per minute per IP.
        The `website` field is a honeypot — leave it empty. A non-empty value silently
        returns a fake success response without creating any URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicCreateURLRequest"
            example:
              url: "https://example.com/very/long/path?with=params"
      responses:
        "201":
          description: Short URL created (or fake decoy response if honeypot triggered)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateURLResponse"
              example:
                short_code: "k7Xm2p"
                short_url: "http://localhost:8080/k7Xm2p"
                original_url: "https://example.com/very/long/path?with=params"
                expires_at: "2025-07-14T00:00:00Z"
                created_at: "2025-06-14T12:00:00Z"
        "400":
          description: Invalid URL format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_url:
                  value:
                    error:
                      code: invalid_url
                      message: The URL format is invalid
                invalid_body:
                  value:
                    error:
                      code: invalid_body
                      message: Request body is invalid JSON
        "422":
          description: URL flagged as unsafe by Safe Browsing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: unsafe_url
                  message: The URL has been flagged as unsafe
        "429":
          description: Too many requests (5 req/min per IP)
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: string
              example: "60"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.
  /api/v1/urls/{code}:
    parameters:
      - name: code
        in: path
        required: true
        description: Short code or custom alias
        schema:
          type: string
        example: k7Xm2p

    get:
      summary: Get URL details
      operationId: getURL
      tags: [URLs]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: URL details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/URL"
              example:
                short_code: k7Xm2p
                short_url: "http://localhost:8080/k7Xm2p"
                original_url: "https://example.com"
                is_custom: false
                expires_at: null
                created_at: "2025-06-01T12:00:00Z"
                click_count: 42
        "404":
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: not_found
                  message: Short URL not found
        "429":
          description: Too many requests
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: string
              example: "60"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.

    patch:
      summary: Update URL expiry
      operationId: updateURL
      tags: [URLs]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateURLRequest"
            example:
              expires_in: "30d"
      responses:
        "200":
          description: Updated URL details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/URL"
              example:
                short_code: k7Xm2p
                short_url: "http://localhost:8080/k7Xm2p"
                original_url: "https://example.com"
                is_custom: false
                expires_at: "2025-07-01T00:00:00Z"
                created_at: "2025-06-01T12:00:00Z"
                click_count: 0
        "400":
          description: Invalid request body or expires_in format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete short URL
      operationId: deleteURL
      tags: [URLs]
      security:
        - ApiKeyAuth: []
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: not_found
                  message: Short URL not found
        "429":
          description: Too many requests
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: string
              example: "60"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.
  /api/v1/urls/{code}/qr:
    parameters:
      - name: code
        in: path
        required: true
        description: Short code or custom alias
        schema:
          type: string
        example: k7Xm2p
      - name: size
        in: query
        required: false
        description: Image size in pixels (min 128, max 1024, default 256)
        schema:
          type: integer
          minimum: 128
          maximum: 1024
          default: 256
        example: 256
      - name: format
        in: query
        required: false
        description: Image format (default png)
        schema:
          type: string
          enum: [png, jpeg, svg]
          default: png
        example: png

    get:
      summary: Get QR code for short URL
      operationId: getQRCode
      tags: [URLs]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: QR code image (PNG, JPEG, or SVG depending on format param)
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/svg+xml:
              schema:
                type: string
        "404":
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: not_found
                  message: Short URL not found
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: unauthorized
                  message: Missing or invalid API key
        "429":
          description: Too many requests
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: string
              example: "60"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.
  /{code}:
    get:
      summary: Redirect to original URL
      operationId: redirect
      tags: [Redirect]
      parameters:
        - name: code
          in: path
          required: true
          description: Short code or custom alias
          schema:
            type: string
          example: k7Xm2p
      responses:
        "302":
          description: Redirect to original URL
          headers:
            Location:
              description: The original URL
              schema:
                type: string
                format: uri
              example: "https://example.com/very/long/path"
        "404":
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: not_found
                  message: Short URL not found
        "410":
          description: Short URL has expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: expired
                  message: This short URL has expired
...
components:
  schemas:
    UpdateURLRequest:
      type: object
      required: [expires_in]
      properties:
        expires_in:
          type: string
          description: |
            New expiration duration (e.g. `1h`, `7d`, `30d`).
            Use `"0"` to remove expiry entirely.
          example: "30d"
...
    PublicCreateURLRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: The long URL to shorten (http/https only, max 2048 chars)
          example: "https://example.com/very/long/path?with=params"
        website:
          type: string
          description: |
            Honeypot field — must be left empty by legitimate clients.
            A non-empty value silently returns a fake success response (bot detection).
    CreateURLRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: The long URL to shorten (http/https only, max 2048 chars)
          example: "https://example.com/very/long/path?with=params"
        custom_alias:
          type: string
          description: |
            Optional custom alias (3–30 chars, `[a-zA-Z0-9-]`, no leading/trailing hyphens).
            If omitted, a short code is generated automatically.
          example: my-link
          minLength: 3
          maxLength: 30
        expires_in:
          type: string
          description: |
            Optional expiry duration. Accepted values: `1h`, `7d`, `30d`, `90d`, `365d`.
            Omit or use `"0"` for no expiry.
          example: 30d
    CreateURLResponse:
      type: object
      required: [short_code, short_url, original_url, created_at]
      properties:
        short_code:
          type: string
          description: The generated or custom short code
          example: k7Xm2p
        short_url:
          type: string
          format: uri
          description: Full short URL (base_url + short_code)
          example: "http://localhost:8080/k7Xm2p"
        original_url:
          type: string
          format: uri
          description: The original long URL
          example: "https://example.com/very/long/path?with=params"
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Expiry timestamp in RFC3339 format, null if no expiry
          example: "2025-07-01T00:00:00Z"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp in RFC3339 format
          example: "2025-06-01T12:00:00Z"
    URL:
      type: object
      required: [short_code, short_url, original_url, is_custom, created_at, click_count]
      properties:
        short_code:
          type: string
          example: k7Xm2p
        short_url:
          type: string
          format: uri
          example: "http://localhost:8080/k7Xm2p"
        original_url:
          type: string
          format: uri
          example: "https://example.com"
        is_custom:
          type: boolean
          description: True if the short code was user-specified
          example: false
        expires_at:
          type: string
          format: date-time
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: "2025-06-01T12:00:00Z"
        click_count:
          type: integer
          format: int64
          description: Number of times this short URL was accessed
          example: 42
    ListURLsResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/URL"
        pagination:
          $ref: "#/components/schemas/Pagination"
    Pagination:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          description: Total number of URLs
          example: 156
        total_pages:
          type: integer
          example: 8
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - invalid_url
            - invalid_expires
            - invalid_body
            - invalid_alias
            - alias_taken
            - reserved_path
            - not_found
            - expired
            - unauthorized
            - rate_limited
            - batch_empty
            - batch_too_large
            - unsafe_url
            - internal_error
          example: invalid_url
        message:
          type: string
          description: Human-readable error description
          example: The URL format is invalid
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (added in Phase 3)