# internal/api/handler.go
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"image/jpeg"
	"log/slog"
	"net/http"
	"strconv"
	"time"

	"github.com/anIcedAntFA/goshort/internal/shortener"
	"github.com/go-chi/chi/v5"
	qrcode "github.com/skip2/go-qrcode"
)

// Handler holds the HTTP handler dependencies.
type Handler struct {
	svc     shortener.Service
	cache   shortener.Cache
	logger  *slog.Logger
	baseURL string
}

// NewHandler creates a new Handler.
func NewHandler(svc shortener.Service, c shortener.Cache, logger *slog.Logger, baseURL string) *Handler {
	return &Handler{svc: svc, cache: c, logger: logger, baseURL: baseURL}
}

type createURLRequest struct {
	URL         string `json:"url"`
	CustomAlias string `json:"custom_alias"`
	ExpiresIn   string `json:"expires_in"`
}

type createURLResponse struct {
	ShortCode   string  `json:"short_code"`
	ShortURL    string  `json:"short_url"`
	OriginalURL string  `json:"original_url"`
	ExpiresAt   *string `json:"expires_at"`
	CreatedAt   string  `json:"created_at"`
	Title       string  `json:"title,omitempty"`
	Description string  `json:"description,omitempty"`
}

type urlResponse struct {
	ShortCode   string  `json:"short_code"`
	ShortURL    string  `json:"short_url"`
	OriginalURL string  `json:"original_url"`
	IsCustom    bool    `json:"is_custom"`
	ExpiresAt   *string `json:"expires_at"`
	CreatedAt   string  `json:"created_at"`
	ClickCount  int64   `json:"click_count"`
	Title       string  `json:"title,omitempty"`
	Description string  `json:"description,omitempty"`
}

type listURLsResponse struct {
	Data       []urlResponse  `json:"data"`
	Pagination paginationMeta `json:"pagination"`
}

type paginationMeta struct {
	Page       int `json:"page"`
	PerPage    int `json:"per_page"`
	Total      int `json:"total"`
	TotalPages int `json:"total_pages"`
}

// CreateURL handles POST /api/v1/urls.
func (h *Handler) CreateURL(w http.ResponseWriter, r *http.Request) {
	var req createURLRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondJSON(w, http.StatusBadRequest, errorResponse{Error: errorDetail{
			Code:    "invalid_body",
			Message: "Request body is invalid JSON",
		}})
		return
	}

	url, err := h.svc.Create(r.Context(), shortener.CreateRequest{
		URL:         req.URL,
		CustomAlias: req.CustomAlias,
		ExpiresIn:   req.ExpiresIn,
	})
	if err != nil {
		respondError(w, err)
		return
	}

	urlType := "generated"
	if url.IsCustom {
		urlType = "custom"
	}
	urlsCreatedTotal.WithLabelValues(urlType).Inc()

	respondJSON(w, http.StatusCreated, createURLResponse{
		ShortCode:   url.ShortCode,
		ShortURL:    fmt.Sprintf("%s/%s", h.baseURL, url.ShortCode),
		OriginalURL: url.OriginalURL,
		ExpiresAt:   formatTimePtr(url.ExpiresAt),
		CreatedAt:   url.CreatedAt.Format(time.RFC3339),
		Title:       url.Title,
		Description: url.Description,
	})
}

// Redirect handles GET /{code} using cache-aside pattern.
func (h *Handler) Redirect(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")

	// Cache hit — skip DB entirely.
	if cached, ok := h.cache.Get(r.Context(), "short:"+code); ok {
		redirectsTotal.WithLabelValues("302").Inc()
		clickCtx := context.WithoutCancel(r.Context())
		go func() {
			if err := h.svc.IncrementClicks(clickCtx, code); err != nil {
				h.logger.Warn("increment clicks failed", "code", code, "error", err)
			}
		}()
		w.Header().Set("Cache-Control", "private, max-age=0, no-cache")
		http.Redirect(w, r, cached, http.StatusFound) //nolint:gosec // URL validated on creation
		return
	}

	url, err := h.svc.GetByCode(r.Context(), code)
	if err != nil {
		if errors.Is(err, shortener.ErrExpired) {
			redirectsTotal.WithLabelValues("410").Inc()
		} else {
			redirectsTotal.WithLabelValues("404").Inc()
		}
		respondError(w, err)
		return
	}

	// Populate cache for subsequent requests.
	_ = h.cache.Set(r.Context(), "short:"+code, url.OriginalURL, cacheTTL(url))

	redirectsTotal.WithLabelValues("302").Inc()

	// Fire-and-forget: detach from request context so cancellation on response send doesn't abort the write.
	clickCtx := context.WithoutCancel(r.Context())
	go func() {
		if err := h.svc.IncrementClicks(clickCtx, code); err != nil {
			h.logger.Warn("increment clicks failed", "code", code, "error", err)
		}
	}()

	w.Header().Set("Cache-Control", "private, max-age=0, no-cache")
	http.Redirect(w, r, url.OriginalURL, http.StatusFound) //nolint:gosec // URL validated on creation
}

// ListURLs handles GET /api/v1/urls.
func (h *Handler) ListURLs(w http.ResponseWriter, r *http.Request) {
	page := parseIntQuery(r, "page", 1)
	perPage := parseIntQuery(r, "per_page", 20)

	urls, total, err := h.svc.List(r.Context(), shortener.ListOptions{
		Page:    page,
		PerPage: perPage,
	})
	if err != nil {
		respondError(w, err)
		return
	}

	data := make([]urlResponse, len(urls))
	for i, u := range urls {
		data[i] = toURLResponse(&u, h.baseURL)
	}

	totalPages := total / perPage
	if total%perPage != 0 {
		totalPages++
	}

	respondJSON(w, http.StatusOK, listURLsResponse{
		Data: data,
		Pagination: paginationMeta{
			Page:       page,
			PerPage:    perPage,
			Total:      total,
			TotalPages: totalPages,
		},
	})
}

// GetURL handles GET /api/v1/urls/{code}.
func (h *Handler) GetURL(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")

	url, err := h.svc.GetByCode(r.Context(), code)
	if err != nil {
		respondError(w, err)
		return
	}

	respondJSON(w, http.StatusOK, toURLResponse(url, h.baseURL))
}

// DeleteURL handles DELETE /api/v1/urls/{code}.
func (h *Handler) DeleteURL(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")

	if err := h.svc.Delete(r.Context(), code); err != nil {
		respondError(w, err)
		return
	}

	_ = h.cache.Delete(r.Context(), "short:"+code)
	w.WriteHeader(http.StatusNoContent)
}

// HealthCheck handles GET /health.
func (h *Handler) HealthCheck(w http.ResponseWriter, _ *http.Request) {
	respondJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

type batchCreateURLRequest struct {
	URLs []createURLRequest `json:"urls"`
}

type batchItemResponse struct {
	ShortCode   string       `json:"short_code,omitempty"`
	ShortURL    string       `json:"short_url,omitempty"`
	OriginalURL string       `json:"original_url,omitempty"`
	ExpiresAt   *string      `json:"expires_at,omitempty"`
	CreatedAt   string       `json:"created_at,omitempty"`
	Error       *errorDetail `json:"error,omitempty"`
}

type batchSummary struct {
	Total   int `json:"total"`
	Success int `json:"success"`
	Failed  int `json:"failed"`
}

type batchCreateURLResponse struct {
	Results []batchItemResponse `json:"results"`
	Summary batchSummary        `json:"summary"`
}

// BatchCreateURL handles POST /api/v1/urls/batch.
func (h *Handler) BatchCreateURL(w http.ResponseWriter, r *http.Request) {
	var req batchCreateURLRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondJSON(w, http.StatusBadRequest, errorResponse{Error: errorDetail{
			Code:    "invalid_body",
			Message: "Request body is invalid JSON",
		}})
		return
	}

	svcReqs := make([]shortener.CreateRequest, len(req.URLs))
	for i, u := range req.URLs {
		svcReqs[i] = shortener.CreateRequest{
			URL:         u.URL,
			CustomAlias: u.CustomAlias,
			ExpiresIn:   u.ExpiresIn,
		}
	}

	batchResults, err := h.svc.CreateBatch(r.Context(), svcReqs)
	if err != nil {
		respondError(w, err)
		return
	}

	items := make([]batchItemResponse, len(batchResults))
	var succeeded int
	for i, br := range batchResults {
		if br.Error != nil {
			items[i] = batchItemResponse{Error: toErrorDetail(br.Error)}
		} else {
			succeeded++
			urlType := "generated"
			if br.URL.IsCustom {
				urlType = "custom"
			}
			urlsCreatedTotal.WithLabelValues(urlType).Inc()
			items[i] = batchItemResponse{
				ShortCode:   br.URL.ShortCode,
				ShortURL:    fmt.Sprintf("%s/%s", h.baseURL, br.URL.ShortCode),
				OriginalURL: br.URL.OriginalURL,
				ExpiresAt:   formatTimePtr(br.URL.ExpiresAt),
				CreatedAt:   br.URL.CreatedAt.Format(time.RFC3339),
			}
		}
	}

	respondJSON(w, http.StatusOK, batchCreateURLResponse{
		Results: items,
		Summary: batchSummary{
			Total:   len(batchResults),
			Success: succeeded,
			Failed:  len(batchResults) - succeeded,
		},
	})
}

type updateURLRequest struct {
	ExpiresIn string `json:"expires_in"`
}

// UpdateURL handles PATCH /api/v1/urls/{code}.
// Only expires_in is mutable; use "0" to remove the expiry.
func (h *Handler) UpdateURL(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")

	var req updateURLRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondJSON(w, http.StatusBadRequest, errorResponse{Error: errorDetail{
			Code:    "invalid_body",
			Message: "Request body is invalid JSON",
		}})
		return
	}

	url, err := h.svc.Update(r.Context(), code, shortener.UpdateRequest{
		ExpiresIn: req.ExpiresIn,
	})
	if err != nil {
		respondError(w, err)
		return
	}

	_ = h.cache.Delete(r.Context(), "short:"+code)
	respondJSON(w, http.StatusOK, toURLResponse(url, h.baseURL))
}

// GetQRCode handles GET /api/v1/urls/{code}/qr.
// Returns a QR code image. Size is clamped to [128, 1024]. Format defaults to PNG.
func (h *Handler) GetQRCode(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")

	size := parseIntQuery(r, "size", 256)
	if size < 128 {
		size = 128
	}
	if size > 1024 {
		size = 1024
	}

	format := r.URL.Query().Get("format")
	if format == "" {
		format = "png"
	}
	switch format {
	case "png", "jpeg", "svg":
	default:
		respondJSON(w, http.StatusBadRequest, errorResponse{Error: errorDetail{
			Code:    "invalid_format",
			Message: "Format must be png, jpeg, or svg",
		}})
		return
	}

	if _, err := h.svc.GetByCode(r.Context(), code); err != nil {
		respondError(w, err)
		return
	}

	shortURL := fmt.Sprintf("%s/%s", h.baseURL, code)
	h.writeQRImage(w, code, shortURL, format, size)
}

func (h *Handler) writeQRImage(w http.ResponseWriter, code, shortURL, format string, size int) {
	switch format {
	case "png":
		data, err := qrcode.Encode(shortURL, qrcode.Medium, size)
		if err != nil {
			h.respondQRError(w, code, err)
			return
		}
		w.Header().Set("Content-Type", "image/png")
		_, _ = w.Write(data) //nolint:errcheck,gosec

	case "jpeg":
		q, err := qrcode.New(shortURL, qrcode.Medium)
		if err != nil {
			h.respondQRError(w, code, err)
			return
		}
		w.Header().Set("Content-Type", "image/jpeg")
		_ = jpeg.Encode(w, q.Image(size), &jpeg.Options{Quality: 90}) //nolint:errcheck,gosec

	case "svg":
		q, err := qrcode.New(shortURL, qrcode.Medium)
		if err != nil {
			h.respondQRError(w, code, err)
			return
		}
		w.Header().Set("Content-Type", "image/svg+xml")
		_, _ = w.Write(qrToSVG(q.Bitmap(), size)) //nolint:errcheck,gosec
	}
}

func qrToSVG(bitmap [][]bool, size int) []byte {
	gridLen := len(bitmap)
	moduleSize := size / gridLen
	totalSize := moduleSize * gridLen
	var buf bytes.Buffer
	fmt.Fprintf(&buf, `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 %d %d">`, totalSize, totalSize)
	buf.WriteString(`<rect width="100%" height="100%" fill="white"/>`)
	for y, row := range bitmap {
		for x, black := range row {
			if black {
				fmt.Fprintf(&buf, `<rect x="%d" y="%d" width="%d" height="%d" fill="black"/>`,
					x*moduleSize, y*moduleSize, moduleSize, moduleSize)
			}
		}
	}
	buf.WriteString(`</svg>`)
	return buf.Bytes()
}

func (h *Handler) respondQRError(w http.ResponseWriter, code string, err error) {
	h.logger.Error("qr code generation failed", "code", code, "error", err)
	respondJSON(w, http.StatusInternalServerError, errorResponse{Error: errorDetail{
		Code:    "internal_error",
		Message: "An internal error occurred",
	}})
}

type publicCreateURLRequest struct {
	URL     string `json:"url"`
	Website string `json:"website"` // honeypot: non-empty means bot
}

// PublicCreateURL handles POST /api/v1/urls/public.
// No auth required. Forces 30-day expiry, no custom alias.
// If the honeypot field "website" is non-empty, returns a fake 201 without storing anything.
func (h *Handler) PublicCreateURL(w http.ResponseWriter, r *http.Request) {
	var req publicCreateURLRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondJSON(w, http.StatusBadRequest, errorResponse{Error: errorDetail{
			Code:    "invalid_body",
			Message: "Request body is invalid JSON",
		}})
		return
	}

	if req.Website != "" {
		expiresAt := time.Now().Add(30 * 24 * time.Hour).Format(time.RFC3339)
		respondJSON(w, http.StatusCreated, createURLResponse{
			ShortCode:   "decoy",
			ShortURL:    fmt.Sprintf("%s/decoy", h.baseURL),
			OriginalURL: req.URL,
			ExpiresAt:   &expiresAt,
			CreatedAt:   time.Now().Format(time.RFC3339),
		})
		return
	}

	url, err := h.svc.Create(r.Context(), shortener.CreateRequest{
		URL:       req.URL,
		ExpiresIn: "30d",
	})
	if err != nil {
		respondError(w, err)
		return
	}

	urlsCreatedTotal.WithLabelValues("generated").Inc()
	respondJSON(w, http.StatusCreated, createURLResponse{
		ShortCode:   url.ShortCode,
		ShortURL:    fmt.Sprintf("%s/%s", h.baseURL, url.ShortCode),
		OriginalURL: url.OriginalURL,
		ExpiresAt:   formatTimePtr(url.ExpiresAt),
		CreatedAt:   url.CreatedAt.Format(time.RFC3339),
		Title:       url.Title,
		Description: url.Description,
	})
}

func parseIntQuery(r *http.Request, key string, defaultVal int) int {
	raw := r.URL.Query().Get(key)
	if raw == "" {
		return defaultVal
	}
	v, err := strconv.Atoi(raw)
	if err != nil || v < 1 {
		return defaultVal
	}
	return v
}

func toURLResponse(u *shortener.URL, baseURL string) urlResponse {
	return urlResponse{
		ShortCode:   u.ShortCode,
		ShortURL:    fmt.Sprintf("%s/%s", baseURL, u.ShortCode),
		OriginalURL: u.OriginalURL,
		IsCustom:    u.IsCustom,
		ExpiresAt:   formatTimePtr(u.ExpiresAt),
		CreatedAt:   u.CreatedAt.Format(time.RFC3339),
		ClickCount:  u.ClickCount,
		Title:       u.Title,
		Description: u.Description,
	}
}

func formatTimePtr(t *time.Time) *string {
	if t == nil {
		return nil
	}
	s := t.Format(time.RFC3339)
	return &s
}

// cacheTTL returns how long to cache a URL.
// The TTL is capped at 24 hours and aligned to the URL's expiry so we never
// serve an expired URL from cache.
func cacheTTL(u *shortener.URL) time.Duration {
	const defaultTTL = 24 * time.Hour
	if u.ExpiresAt == nil {
		return defaultTTL
	}
	remaining := time.Until(*u.ExpiresAt)
	if remaining <= 0 {
		return 0
	}
	if remaining > defaultTTL {
		return defaultTTL
	}
	return remaining
}